requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
python-multipart
pydantic[email]
twilio
python-dotenv
